benchmarks: move plotting to benchmarks/plotting.py + text_auto + hover absolutes
- New ``benchmarks/plotting.py`` module owns the three views
  (``plot_compare`` / ``plot_sweep`` / ``plot_scaling``) plus a
  ``RENDERERS`` dispatch dict. cli.py drops ~140 lines and just imports
  ``PlotView`` + ``RENDERERS``; plotly is still lazy-loaded inside the
  view functions so importing the module without plotly works.

- ``compare`` bar chart and ``sweep`` heatmap now use ``text_auto``
  so values render inside each bar / cell.

- Hover info upgraded:
  - compare hover shows the per-test median of *both* snapshots
    (formatted to 4 significant figures) in addition to the delta %.
  - sweep hover shows the absolute median (s) alongside the ratio, via
    a customdata + hovertemplate plumbed through ``update_traces``.

scaling view already shows the absolute median on hover by virtue of
being a line chart.
